move lowering into printers/codegen.h
