+++ symengine/CMakeLists.txt
@@ -37,7 +37,6 @@ add_subdirectory(utilities/matchpycpp)
 set(SRC
     add.cpp
     basic.cpp
-    codegen_lowering.cpp
     complex.cpp
     complex_double.cpp
     constants.cpp
@@ -169,7 +168,6 @@ set(HEADERS
     complex_double.h
     complex.h
     complex_mpc.h
-    codegen_lowering.h
     constants.h
     cwrapper.h
     derivative.h
